+++ frontend/package.json
@@ -2,9 +2,9 @@
 	"name": "windmill",
 	"version": "1.17.1",
 	"scripts": {
-		"dev": "svelte-kit dev",
-		"build": "svelte-kit build",
-		"preview": "svelte-kit preview",
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
 		"check": "svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
@@ -16,7 +16,6 @@
 	},
 	"devDependencies": {
 		"@playwright/test": "^1.23.2",
-		"@sveltejs/adapter-node": "^1.0.0-next.79",
 		"@sveltejs/adapter-static": "^1.0.0-next.35",
 		"@sveltejs/kit": "next",
 		"@tailwindcss/forms": "^0.5.1",
@@ -48,7 +47,8 @@
 		"svelte2tsx": "^0.5.11",
 		"tailwindcss": "^3.1.6",
 		"tslib": "^2.3.1",
-		"typescript": "^4.7.4"
+		"typescript": "^4.7.4",
+		"vite": "^2.9.14"
 	},
 	"type": "module",
 	"dependencies": {
