+++ frontend/svelte.config.js
@@ -1,13 +1,6 @@
 import preprocess from 'svelte-preprocess'
 import adapter from '@sveltejs/adapter-static'
-const ppath = process.env.PREVIEW_PATH
 
-import { readFileSync } from 'fs'
-import { fileURLToPath } from 'url'
-
-const file = fileURLToPath(new URL('package.json', import.meta.url))
-const json = readFileSync(file, 'utf8')
-const version = JSON.parse(json)
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
@@ -20,23 +13,6 @@ const config = {
 	],
 
 	kit: {
-		paths: {
-			base: ppath ? ppath : ''
-		},
-		vite: {
-			define: {
-				__pkg__: version
-			},
-			ssr: {
-				noExternal: ['dayjs']
-			},
-			optimizeDeps: {
-				include: [
-					'highlight.js',
-					'highlight.js/lib/core',
-				]
-			},
-		},
 		adapter: adapter({
 			// default options are shown
 			pages: 'build',
@@ -44,7 +20,7 @@ const config = {
 			fallback: '200.html'
 		}),
 		prerender: {
-			default: false
+			default: false,
 		},
 	}
 }
